refactor: 💡 change submit to update-status
