+++ src/features/update-status/components/news-table.tsx
@@ -2,7 +2,7 @@
 import { getAllNews } from "@/apis/prisma/fetch-news";
 import { StatusCodeView } from "@/components/status-code-view";
 import type { NewsContext } from "@/features/dump/stores/news-context";
-import { DumpTable } from "@/features/submit/components/dump-table";
+import { DumpTable } from "@/features/update-status/components/dump-table";
 
 export async function NewsTable() {
 	try {
